+++ src/index.ts
@@ -1,4 +1,4 @@
-export { init, type SentryConfig } from "./instrumentation.js"
+export { init } from "./instrumentation.js"
 export { getLoadContext, type GetLoadContextOptions } from "./load_context.js"
 export { default as logger } from "./logger.js"
 export { startMetrics } from "./metrics.js"